exam-coach/backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
sqlalchemy==2.0.21
pydantic==2.11.1
openai==1.68.2
python-dotenv==1.0.0
aiofiles==23.1.0
anyio==3.7.1
starlette==0.27.0
